RFC 6979 and `0.5.0` prep (#63)
* version3: Implement signing/verification with RFC6979 k-values

* version3: Clean up documentation leftovers from when we used ring

* Rename Error::Base64Decoding -> Error::Base64

* Move keys and their version definition to the respective version modules

* Fix imports

* Feature-gate PASERK impls based on versions

* Feature-gate Footer tests

* Feature-gate UntrustedToken tests

* Fix various imports

* Enable test jobs for select versions enabled only

* Update fuzzing targets

* 0.5.0-alpha.7

* version3: Now that we use deterministic nonces, test for correctly reproduced sigs as well

* No default features = no doctests
