+++ fuzz/Cargo.toml
@@ -13,7 +13,6 @@ libfuzzer-sys = "0.3"
 rand_chacha = "0.2.2"
 rand_core = "0.5.0"
 ed25519-compact = "1.0.11"
-ring = "0.16.20"
 
 [dependencies.pasetors]
 path = ".."
@@ -26,12 +25,6 @@ members = ["."]
 [[bin]]
 name = "fuzz_pasetors"
 path = "fuzz_targets/fuzz_pasetors.rs"
-test = false
-doc = false
-
-[[bin]]
-name = "fuzz_point_comp"
-path = "fuzz_targets/fuzz_point_comp.rs"
 test = false
 doc = false
